feat: add custom hook useProductDetail
