+++ src/components/product/ProductDescription.module.css
@@ -1,18 +1,19 @@
-.description{
-    margin-bottom: 1rem;
-    padding: 1rem;
-    border: solid black 1px;
-    background-color: var(--green-100);
-    border-radius: 2%;
+.description {
+  margin-bottom: 1rem;
+  padding: 1rem;
+  border: solid black 1px;
+  background-color: var(--green-100);
+  border-radius: 2%;
 }
 
-.description h3{
-    font-size: 2rem;
-    text-decoration: underline;
-    text-align: center;
+.description h3 {
+  font-size: 2rem;
+  text-decoration: underline;
+  text-align: center;
 }
 
-.description li{
-    text-transform: capitalize;
-    list-style: none;
+.description li {
+  text-transform: capitalize;
+  list-style: none;
+  margin-top: 0.3rem;
 }
